+++ src/features/widget/volumeMixer.ts
@@ -17,6 +17,7 @@ import { updateMenuSeparators } from "../../libs/shell/quickSettingsUtils.js"
 import Maid from "../../libs/shared/maid.js"
 import Global from "../../global.js"
 import Logger from "../../libs/shared/logger.js"
+import { VerticalProp } from "../../libs/shell/compat.js"
 
 const ALLOW_AMPLIFIED_VOLUME_KEY = 'allow-volume-above-100-percent'
 
@@ -255,7 +256,7 @@ class VolumeMixerItem extends St.BoxLayout {
 		options: VolumeMixerItem.Options
 	) {
 		super._init({
-			vertical: true,
+			...VerticalProp,
 			style_class: "QSTWEAKS-item",
 		})
 		this._control = control
@@ -330,9 +331,9 @@ class VolumeMixerList extends St.BoxLayout {
 
 	constructor(options: VolumeMixerList.Options) {
 		super({
+			...VerticalProp,
 			style_class: "QSTWEAKS-volume-mixer",
 			clip_to_allocation: true,
-			vertical: true,
 			x_expand: true,
 		})
 		this._options = options
@@ -467,7 +468,7 @@ class VolumeMixerWidget extends St.BoxLayout {
 	}
 	_init(options: VolumeMixerWidget.Options) {
 		super._init({
-			vertical: true,
+			...VerticalProp,
 		} as Partial<St.BoxLayout.ConstructorProps>)
 
 		this._options = options
@@ -500,7 +501,7 @@ class VolumeMixerWidget extends St.BoxLayout {
 	// Scroll view
 	_createScroll() {
 		this._sections = new St.BoxLayout({
-			vertical: true,
+			...VerticalProp,
 			x_expand: true,
 			y_expand: true,
 		})